+++ localflavor/us/models.py
@@ -1,6 +1,7 @@
 from django.utils.translation import ugettext_lazy as _
 from django.db.models.fields import CharField
 
+from . import forms
 from .us_states import STATE_CHOICES, USPS_CHOICES
 
 
@@ -19,9 +20,14 @@ class USStateField(CharField):
 
 class USPostalCodeField(CharField):
     """"
-    A model field that forms represent as a ``forms.USPSSelect`` field
-    and stores the two-letter U.S Postal Service abbreviation in the
-    database.
+    A model field that forms represent as a
+    :class:`~localflavor.us.forms.USPSSelect`` field and stores the two-letter
+    U.S. Postal Service abbreviation in the database.
+
+    .. note::
+
+        If you are looking for a model field that validates U.S. ZIP codes
+        please use :class:`~localflavor.us.models.USZipCodeField`.
     """
     description = _("U.S. postal code (two uppercase letters)")
 
@@ -29,6 +35,32 @@ class USPostalCodeField(CharField):
         kwargs['choices'] = USPS_CHOICES
         kwargs['max_length'] = 2
         super(USPostalCodeField, self).__init__(*args, **kwargs)
+
+
+class USZipCodeField(CharField):
+    """
+    A model field that forms represent as a
+    :class:`~localflavor.us.forms.USZipCodeField` field and stores the
+    U.S. ZIP code in the database.
+
+    .. note::
+
+        If you are looking for a model field with a list of U.S. Postal Service
+        locations please use :class:`~localflavor.us.models.USPostalCodeField`.
+
+    .. versionadded:: 1.1
+
+    """
+    description = _("U.S. ZIP code")
+
+    def __init__(self, *args, **kwargs):
+        kwargs['max_length'] = 10
+        super(USZipCodeField, self).__init__(*args, **kwargs)
+
+    def formfield(self, **kwargs):
+        defaults = {'form_class': forms.USZipCodeField}
+        defaults.update(kwargs)
+        return super(USZipCodeField, self).formfield(**defaults)
 
 
 class PhoneNumberField(CharField):