+++ tests/test_us/forms.py
@@ -10,4 +10,4 @@ class USPlaceForm(ModelForm):
     class Meta:
         model = USPlace
         fields = ('state', 'state_req', 'state_default', 'postal_code', 'name',
-                  'ssn')
+                  'ssn', 'zip_code')