+++ tests/test_us/models.py
@@ -1,7 +1,8 @@
 from django.db import models
 
 from localflavor.us.models import (USStateField, USPostalCodeField,
-                                   USSocialSecurityNumberField)
+                                   USSocialSecurityNumberField,
+                                   USZipCodeField)
 
 
 class USPlace(models.Model):
@@ -11,3 +12,4 @@ class USPlace(models.Model):
     postal_code = USPostalCodeField(blank=True)
     name = models.CharField(max_length=20)
     ssn = USSocialSecurityNumberField(blank=True)
+    zip_code = USZipCodeField(blank=True)